Time: 744 ms (28.97%), Space: 180.2 MB (74.93%) - LeetHub

+++ 1457-pseudo-palindromic-paths-in-a-binary-tree/1457-pseudo-palindromic-paths-in-a-binary-tree.cpp
@@ -11,7 +11,7 @@
  */
 class Solution {
 public:
-    long long  ans,c, x;
+    int  ans,c, x;
     void GetAns(TreeNode* node){
         if(node==nullptr)
            return;
